Only backup when we ask for it. Check that items have a valid view set for the class. Better logging for invalid views and classes.

+++ startup/cron.php
@@ -34,6 +34,9 @@ function checkConsistency()
         $versions = $variant->getVersions();
         foreach ($versions as $version)
         {
+        	if (!$version->getClass()->isValidView($version->getView()))
+        		$log->warn('Itemversion '.$version->getId().' has invalid view '.$version->getView()->getId().' for class '.$version->getClass()->getId().'.');
+
         	$path = $version->getStoragePath();
           $results = $_STORAGE->query('SELECT * FROM File WHERE itemversion='.$version->getId().';');
           while ($results->valid())
@@ -128,19 +131,22 @@ if ($output !== FALSE)
 	fclose($output);
 }
 
-if ((is_executable($_PREFS->getPref('tools.tar'))) && (is_executable($_PREFS->getPref('tools.mysqldump'))))
+if (isset($_GET['backup']))
 {
-	$backupfile = $_PREFS->getPref('storage.backup').'/'.$_PREFS->getPref('url.host.1.hostname').'-';
-	$backupfile .= date('Ymd-Hi');
-	$backupfile .= '.tar.bz';
-	system($_PREFS->getPref('tools.mysqldump').' --result-file='.$_PREFS->getPref('storage.site').'/database.sql --add-drop-table --ignore-table='.$_PREFS->getPref('storage.mysql.database').'.Keywords -u '.$_PREFS->getPref('storage.mysql.user').' -p'.$_PREFS->getPref('storage.mysql.pass').' -e '.$_PREFS->getPref('storage.mysql.database'));
-	system($_PREFS->getPref('tools.tar').' -cjf '.$backupfile.' -C '.$_PREFS->getPref('storage.site').' database.sql files');
-	unlink($_PREFS->getPref('storage.site').'/database.sql');
-}
-else
-{
-	$log = Loggermanager::getLogger('swim.backup');
-	$log->error('Unable to backup since tools are unavailable.');
+	if ((is_executable($_PREFS->getPref('tools.tar'))) && (is_executable($_PREFS->getPref('tools.mysqldump'))))
+	{
+		$backupfile = $_PREFS->getPref('storage.backup').'/'.$_PREFS->getPref('url.host.1.hostname').'-';
+		$backupfile .= date('Ymd-Hi');
+		$backupfile .= '.tar.bz';
+		system($_PREFS->getPref('tools.mysqldump').' --result-file='.$_PREFS->getPref('storage.site').'/database.sql --add-drop-table --ignore-table='.$_PREFS->getPref('storage.mysql.database').'.Keywords -u '.$_PREFS->getPref('storage.mysql.user').' -p'.$_PREFS->getPref('storage.mysql.pass').' -e '.$_PREFS->getPref('storage.mysql.database'));
+		system($_PREFS->getPref('tools.tar').' -cjf '.$backupfile.' -C '.$_PREFS->getPref('storage.site').' database.sql files');
+		unlink($_PREFS->getPref('storage.site').'/database.sql');
+	}
+	else
+	{
+		$log = Loggermanager::getLogger('swim.backup');
+		$log->error('Unable to backup since tools are unavailable.');
+	}
 }
 
 ?>
